Improves the synchronize job and service
